Homework-2 - Copy/requirements.txt
flask
gunicorn
Flask-Failsafe
mysql-connector-python
